requirements.txt
click ~=7.1.2
google-api-python-client ~=1.12.2
google-auth-oauthlib ~=0.4.1
prometheus_client